refactor(rollout): split In Situ into Config + In Progress buckets
Split the single "In Situ" story bucket into two so launch surfaces and
live surfaces read separately:

- In Situ/Config: the Bulk actions story (launch a process from the
  selection bar).
- In Situ/In Progress: Header pill, Firmware settings, Energy, Activity.

Also, per review:
- Drop the Activity active-now and Activity rollout-detail stories.
- Firmware settings page now renders the real settings subnav (the
  shipped SecondaryNavigation), matching SettingsLayout chrome.
- Reboot bulk modal omits the "Apply to" section — scope is already
  fixed by the selection. RolloutConfigModal.scopeTargets is now
  optional; the Apply-to section renders only when provided.
- Bulk Curtail opens the shipped full-screen CurtailmentStartModal
  instead of the generic config modal.
- Rollout strategy option labels drop "Update" ("All at once",
  "In batches").

Story-only + additive framework tweaks; no shipped component or its
onConfirm contract is edited. tsc + eslint clean.
